backend/requirements.txt
fastapi
uvicorn
boto3
pydantic
python-multipart
requests
python-dotenv
